Final Load Test

+++ src/utils/game.ts
@@ -228,10 +228,13 @@ export async function propagateGameChange(interaction: Interaction, game: Game) 
         await channel.send({ content: `Game ${game.id} has been updated by <@${interaction.user.id}>. ${winnerText}`, embeds: [ embed ] });
         await modChannel.send({ content: `Game ${game.id} has been updated by <@${interaction.user.id}>.`, embeds: [ embed ], components: [ components ] });
         await reply(interaction, { content: `${games.length} game ${games.length == 1 ? "update was" : "updates were"} applied successfully.` });
-        if (interaction.message?.deletable) {
-            try {
-                await interaction.message?.delete()
-            } catch (_) { }
+
+        if (interaction.isButton() || interaction.isModalSubmit()) {
+            if (interaction.message?.deletable) {
+                try {
+                    await interaction.message?.delete()
+                } catch (_) { }
+            }
         }
     }
 }
